Import router tests now override settings directly
Previously we were using monkeypatch to set the environment variables. But even with the context, these were persisting in the settings maintained by other routers past the clean-up in the tests. That meant that other router tests were being run in a `production` environment depending on test ordering. We now temporarily override the imports settings object, then reinialize it at the end of the test.