Split Annotation into separate types
In order to make annotations more fit for purpose I've split them into
multiple types to reflect the different annotations we might expect.

Initially I've added a `BooleanAnnotation` which supports the explicit
application or non-application of a label within a scheme. This is most
likely to be used with our classifiers.

There is also a `ScoreAnnotation` which looks very similar to a
`BooleanAnnotation` but lacks the binary classification element. It just
attributes a score with a label within a scheme.